# backend/swagger.json
{
  "openapi": "3.0.3",
  "paths": {
...
    "/trade/spot/limit": {
      "post": {
        "description": "Execute a limit order for spot trading",
        "operationId": "TradeController_executeSpotLimitOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LimitOrderSpotRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Spot limit order executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderExecutionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Execute spot limit order",
        "tags": [
          "Trading"
        ]
      }
    },
    "/trade/spot/market": {
      "post": {
        "description": "Execute a market order for spot trading",
        "operationId": "TradeController_executeSpotMarketOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketOrderSpotRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Spot market order executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderExecutionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Execute spot market order",
        "tags": [
          "Trading"
        ]
      }
    },
    "/trade/spot/twap": {
      "post": {
        "description": "Execute a TWAP order for spot trading",
        "operationId": "TradeController_executeSpotTwapOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "asset",
                  "side",
                  "usdcSize",
                  "twapRunningTime",
                  "twapRandomize"
                ],
                "properties": {
                  "asset": {
                    "type": "string",
                    "description": "Trading asset",
                    "example": "ETH"
                  },
                  "side": {
                    "type": "string",
                    "enum": [
                      "BUY",
                      "SELL"
                    ],
                    "description": "Trade side",
                    "example": "BUY"
                  },
                  "usdcSize": {
                    "type": "number",
                    "description": "Size in USDC",
                    "example": 1000
                  },
                  "twapRunningTime": {
                    "type": "number",
                    "description": "TWAP running time in minutes",
                    "example": 60
                  },
                  "twapRandomize": {
                    "type": "boolean",
                    "description": "Whether to randomize TWAP",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Spot TWAP order executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderExecutionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Execute spot TWAP order",
        "tags": [
          "Trading"
        ]
      }
    },
    "/trade/perpetual/limit": {
      "post": {
        "description": "Execute a limit order for perpetual trading",
        "operationId": "TradeController_executePerpetualLimitOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LimitOrderPerpetualRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Perpetual limit order executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderExecutionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Execute perpetual limit order",
        "tags": [
          "Trading"
        ]
      }
    },
    "/trade/perpetual/market": {
      "post": {
        "description": "Execute a market order for perpetual trading",
        "operationId": "TradeController_executePerpetualMarketOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketOrderPerpetualRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Perpetual market order executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderExecutionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Execute perpetual market order",
        "tags": [
          "Trading"
        ]
      }
    },
    "/trade/perpetual/twap": {
      "post": {
        "description": "Execute a TWAP order for perpetual trading",
        "operationId": "TradeController_executePerpetualTwapOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "asset",
                  "side",
                  "usdcSize",
                  "leverage",
                  "twapRunningTime",
                  "twapRandomize"
                ],
                "properties": {
                  "asset": {
                    "type": "string",
                    "description": "Trading asset",
                    "example": "ETH"
                  },
                  "side": {
                    "type": "string",
                    "enum": [
                      "BUY",
                      "SELL"
                    ],
                    "description": "Trade side",
                    "example": "BUY"
                  },
                  "usdcSize": {
                    "type": "number",
                    "description": "Size in USDC",
                    "example": 1000
                  },
                  "leverage": {
                    "type": "number",
                    "description": "Leverage amount",
                    "example": 5
                  },
                  "twapRunningTime": {
                    "type": "number",
                    "description": "TWAP running time in minutes",
                    "example": 60
                  },
                  "twapRandomize": {
                    "type": "boolean",
                    "description": "Whether to randomize TWAP",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Perpetual TWAP order executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderExecutionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Execute perpetual TWAP order",
        "tags": [
          "Trading"
        ]
      }
    },
...
  },
  "info": {
    "title": "Brother Terminal Trading API",
    "description": "API for managing trading positions and orders on the Brother Terminal platform for Hyperliquid Engine.",
    "version": "3.0.0",
    "contact": {
      "name": "Brother Terminal",
      "url": "https://brother-terminal.xyz",
      "email": "dev@brother-terminal.xyz"
    }
  },
...
  "servers": [
    {
      "url": "https://api.brother-terminal.xyz",
      "description": "Production (Mainnet)"
    },
    {
      "url": "https://api-testnet.brother-terminal.xyz",
      "description": "Testnet"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "LimitOrderSpotRequestDto": {
        "type": "object",
        "properties": {
          "asset": {
            "type": "string",
            "description": "Trading asset",
            "example": "ETH"
          },
          "side": {
            "type": "string",
            "description": "Trade side",
            "enum": [
              "BUY",
              "SELL"
            ],
            "example": "BUY"
          },
          "usdcSize": {
            "type": "number",
            "description": "Size in USDC",
            "example": 1000
          },
          "price": {
            "type": "number",
            "description": "Limit price for the order",
            "example": 3500.5
          }
        },
        "required": [
          "asset",
          "side",
          "usdcSize",
          "price"
        ]
      },
      "OrderExecutionResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the order was successful",
            "example": true
          },
          "externalTxnHash": {
            "type": "string",
            "description": "External transaction hash if successful",
            "example": "0x1234567890abcdef..."
          },
          "error": {
            "type": "string",
            "description": "Error message if unsuccessful",
            "example": "Insufficient balance"
          }
        },
        "required": [
          "success"
        ]
      },
      "MarketOrderSpotRequestDto": {
        "type": "object",
        "properties": {
          "asset": {
            "type": "string",
            "description": "Trading asset",
            "example": "ETH"
          },
          "side": {
            "type": "string",
            "description": "Trade side",
            "enum": [
              "BUY",
              "SELL"
            ],
            "example": "BUY"
          },
          "usdcSize": {
            "type": "number",
            "description": "Size in USDC",
            "example": 1000
          }
        },
        "required": [
          "asset",
          "side",
          "usdcSize"
        ]
      },
      "LimitOrderPerpetualRequestDto": {
        "type": "object",
        "properties": {
          "asset": {
            "type": "string",
            "description": "Trading asset",
            "example": "ETH"
          },
          "side": {
            "type": "string",
            "description": "Trade side",
            "enum": [
              "BUY",
              "SELL"
            ],
            "example": "BUY"
          },
          "usdcSize": {
            "type": "number",
            "description": "Size in USDC",
            "example": 1000
          },
          "leverage": {
            "type": "number",
            "description": "Leverage amount",
            "example": 5
          },
          "price": {
            "type": "number",
            "description": "Limit price for the order",
            "example": 3500.5
          }
        },
        "required": [
          "asset",
          "side",
          "usdcSize",
          "leverage",
          "price"
        ]
      },
      "MarketOrderPerpetualRequestDto": {
        "type": "object",
        "properties": {
          "asset": {
            "type": "string",
            "description": "Trading asset",
            "example": "ETH"
          },
          "side": {
            "type": "string",
            "description": "Trade side",
            "enum": [
              "BUY",
              "SELL"
            ],
            "example": "BUY"
          },
          "usdcSize": {
            "type": "number",
            "description": "Size in USDC",
            "example": 1000
          },
          "leverage": {
            "type": "number",
            "description": "Leverage amount",
            "example": 5
          }
        },
        "required": [
          "asset",
          "side",
          "usdcSize",
          "leverage"
        ]
      },
...
    }
  }
}